+++ packages/graph/hash_graph/lib/graph/src/api/rest/entity.rs
@@ -16,16 +16,18 @@ use utoipa::{OpenApi, ToSchema};
 use crate::{
     api::rest::{api_resource::RoutedResource, read_from_store, report_to_status_code},
     knowledge::{
-        Entity, EntityId, EntityRootedSubgraph, PersistedEntity, PersistedEntityIdentifier,
-        PersistedEntityMetadata,
+        Entity, EntityId, PersistedEntity, PersistedEntityIdentifier, PersistedEntityMetadata,
     },
-    ontology::AccountId,
+    provenance::{CreatedById, OwnedById, UpdatedById},
+    shared::identifier::GraphElementIdentifier,
     store::{
         error::{EntityDoesNotExist, QueryError},
         query::Expression,
         EntityStore, StorePool,
     },
-    subgraph::StructuralQuery,
+    subgraph::{
+        EdgeKind, Edges, GraphResolveDepths, OutwardEdge, StructuralQuery, Subgraph, Vertex,
+    },
 };
 
 #[derive(OpenApi)]
@@ -39,6 +41,9 @@ use crate::{
     ),
     components(
         schemas(
+            OwnedById,
+            CreatedById,
+            UpdatedById,
             CreateEntityRequest,
             UpdateEntityRequest,
             EntityId,
@@ -47,7 +52,13 @@ use crate::{
             PersistedEntity,
             Entity,
             StructuralQuery,
-            EntityRootedSubgraph,
+            GraphElementIdentifier,
+            Vertex,
+            EdgeKind,
+            OutwardEdge,
+            GraphResolveDepths,
+            Edges,
+            Subgraph,
         )
     ),
     tags(
@@ -81,9 +92,9 @@ struct CreateEntityRequest {
     entity: Entity,
     #[schema(value_type = String)]
     entity_type_id: VersionedUri,
-    owned_by_id: AccountId,
+    owned_by_id: OwnedById,
     entity_id: Option<EntityId>,
-    actor_id: AccountId,
+    actor_id: CreatedById,
 }
 
 #[utoipa::path(
@@ -135,8 +146,7 @@ async fn create_entity<P: StorePool + Send>(
     request_body = StructuralQuery,
     tag = "Entity",
     responses(
-        (status = 200, content_type = "application/json", body = [EntityRootedSubgraph], description = "A list of subgraphs rooted at entities that satisfy the given query, each resolved to the requested depth."),
-
+        (status = 200, content_type = "application/json", body = Subgraph, description = "A subgraph rooted at entities that satisfy the given query, each resolved to the requested depth."),
         (status = 422, content_type = "text/plain", description = "Provided query is invalid"),
         (status = 500, description = "Store error occurred"),
     )
@@ -144,7 +154,7 @@ async fn create_entity<P: StorePool + Send>(
 async fn get_entities_by_query<P: StorePool + Send>(
     pool: Extension<Arc<P>>,
     Json(query): Json<StructuralQuery>,
-) -> Result<Json<Vec<EntityRootedSubgraph>>, StatusCode> {
+) -> Result<Json<Subgraph>, StatusCode> {
     pool.acquire()
         .map_err(|error| {
             tracing::error!(?error, "Could not acquire access to the store");
@@ -210,7 +220,7 @@ struct UpdateEntityRequest {
     entity_id: EntityId,
     #[schema(value_type = String)]
     entity_type_id: VersionedUri,
-    actor_id: AccountId,
+    actor_id: UpdatedById,
 }
 
 #[utoipa::path(