+++ packages/graph/hash_graph/lib/graph/src/store/postgres/ontology/data_type/mod.rs
@@ -7,16 +7,15 @@ use tokio_postgres::GenericClient;
 use type_system::{uri::VersionedUri, DataType};
 
 use crate::{
-    ontology::{AccountId, PersistedDataType, PersistedOntologyMetadata},
+    ontology::{PersistedDataType, PersistedOntologyMetadata},
+    provenance::{CreatedById, OwnedById, UpdatedById},
+    shared::identifier::GraphElementIdentifier,
     store::{
         crud::Read,
-        postgres::{
-            context::PostgresContext, DependencyContext, DependencyContextRef,
-            PersistedOntologyType,
-        },
+        postgres::{context::PostgresContext, DependencyContext, DependencyContextRef},
         AsClient, DataTypeStore, InsertionError, PostgresStore, QueryError, UpdateError,
     },
-    subgraph::{GraphElementIdentifier, StructuralQuery, Subgraph},
+    subgraph::{StructuralQuery, Subgraph},
 };
 
 impl<C: AsClient> PostgresStore<C> {
@@ -39,7 +38,7 @@ impl<C: AsClient> PostgresStore<C> {
                 data_type_id,
                 Some(graph_resolve_depths.data_type_resolve_depth),
                 || async {
-                    Ok(PersistedDataType::from_record(
+                    Ok(PersistedDataType::from(
                         self.read_versioned_ontology_type(data_type_id).await?,
                     ))
                 },
@@ -55,8 +54,8 @@ impl<C: AsClient> DataTypeStore for PostgresStore<C> {
     async fn create_data_type(
         &mut self,
         data_type: DataType,
-        owned_by_id: AccountId,
-        created_by_id: AccountId,
+        owned_by_id: OwnedById,
+        created_by_id: CreatedById,
     ) -> Result<PersistedOntologyMetadata, InsertionError> {
         let transaction = PostgresStore::new(
             self.as_mut_client()
@@ -114,7 +113,7 @@ impl<C: AsClient> DataTypeStore for PostgresStore<C> {
     async fn update_data_type(
         &mut self,
         data_type: DataType,
-        updated_by_id: AccountId,
+        updated_by_id: UpdatedById,
     ) -> Result<PersistedOntologyMetadata, UpdateError> {
         let transaction = PostgresStore::new(
             self.as_mut_client()