+++ packages/graph/hash_graph/tests/integration/postgres/mod.rs
@@ -9,11 +9,14 @@ use std::str::FromStr;
 
 use error_stack::{Report, Result};
 use graph::{
+    identifier::AccountId,
     knowledge::{Entity, EntityId, Link, PersistedEntity, PersistedEntityMetadata, PersistedLink},
     ontology::{
-        AccountId, PersistedDataType, PersistedEntityType, PersistedLinkType,
-        PersistedOntologyMetadata, PersistedPropertyType,
+        PersistedDataType, PersistedEntityType, PersistedLinkType, PersistedOntologyMetadata,
+        PersistedPropertyType,
     },
+    provenance::{CreatedById, OwnedById, RemovedById, UpdatedById},
+    shared::identifier::GraphElementIdentifier,
     store::{
         error::LinkRemovalError,
         query::{Expression, Literal, Path, PathSegment},
@@ -21,9 +24,12 @@ use graph::{
         EntityTypeStore, InsertionError, LinkStore, LinkTypeStore, PostgresStore,
         PostgresStorePool, PropertyTypeStore, QueryError, StorePool, UpdateError,
     },
-    subgraph::{GraphElementIdentifier, GraphResolveDepths, StructuralQuery, Vertex},
+    subgraph::{GraphResolveDepths, StructuralQuery, Vertex},
 };
-use tokio_postgres::{NoTls, Transaction};
+use tokio_postgres::{
+    types::{FromSql, ToSql},
+    NoTls, Transaction,
+};
 use type_system::{uri::VersionedUri, DataType, EntityType, LinkType, PropertyType};
 use uuid::Uuid;
 
@@ -100,8 +106,8 @@ impl DatabaseTestWrapper {
             store
                 .create_data_type(
                     DataType::from_str(data_type).expect("could not parse data type"),
-                    account_id,
-                    account_id,
+                    OwnedById::new(account_id),
+                    CreatedById::new(account_id),
                 )
                 .await?;
         }
@@ -110,8 +116,8 @@ impl DatabaseTestWrapper {
             store
                 .create_property_type(
                     PropertyType::from_str(property_type).expect("could not parse property type"),
-                    account_id,
-                    account_id,
+                    OwnedById::new(account_id),
+                    CreatedById::new(account_id),
                 )
                 .await?;
         }
@@ -121,8 +127,8 @@ impl DatabaseTestWrapper {
             store
                 .create_link_type(
                     LinkType::from_str(link_type).expect("could not parse link type"),
-                    account_id,
-                    account_id,
+                    OwnedById::new(account_id),
+                    CreatedById::new(account_id),
                 )
                 .await?;
         }
@@ -131,8 +137,8 @@ impl DatabaseTestWrapper {
             store
                 .create_entity_type(
                     EntityType::from_str(entity_type).expect("could not parse entity type"),
-                    account_id,
-                    account_id,
+                    OwnedById::new(account_id),
+                    CreatedById::new(account_id),
                 )
                 .await?;
         }
@@ -148,7 +154,11 @@ impl DatabaseApi<'_> {
         data_type: DataType,
     ) -> Result<PersistedOntologyMetadata, InsertionError> {
         self.store
-            .create_data_type(data_type, self.account_id, self.account_id)
+            .create_data_type(
+                data_type,
+                OwnedById::new(self.account_id),
+                CreatedById::new(self.account_id),
+            )
             .await
     }
 
@@ -178,7 +188,7 @@ impl DatabaseApi<'_> {
         data_type: DataType,
     ) -> Result<PersistedOntologyMetadata, UpdateError> {
         self.store
-            .update_data_type(data_type, self.account_id)
+            .update_data_type(data_type, UpdatedById::new(self.account_id))
             .await
     }
 
@@ -187,7 +197,11 @@ impl DatabaseApi<'_> {
         property_type: PropertyType,
     ) -> Result<PersistedOntologyMetadata, InsertionError> {
         self.store
-            .create_property_type(property_type, self.account_id, self.account_id)
+            .create_property_type(
+                property_type,
+                OwnedById::new(self.account_id),
+                CreatedById::new(self.account_id),
+            )
             .await
     }
 
@@ -217,7 +231,7 @@ impl DatabaseApi<'_> {
         property_type: PropertyType,
     ) -> Result<PersistedOntologyMetadata, UpdateError> {
         self.store
-            .update_property_type(property_type, self.account_id)
+            .update_property_type(property_type, UpdatedById::new(self.account_id))
             .await
     }
 
@@ -226,7 +240,11 @@ impl DatabaseApi<'_> {
         entity_type: EntityType,
     ) -> Result<PersistedOntologyMetadata, InsertionError> {
         self.store
-            .create_entity_type(entity_type, self.account_id, self.account_id)
+            .create_entity_type(
+                entity_type,
+                OwnedById::new(self.account_id),
+                CreatedById::new(self.account_id),
+            )
             .await
     }
 
@@ -256,7 +274,7 @@ impl DatabaseApi<'_> {
         entity_type: EntityType,
     ) -> Result<PersistedOntologyMetadata, UpdateError> {
         self.store
-            .update_entity_type(entity_type, self.account_id)
+            .update_entity_type(entity_type, UpdatedById::new(self.account_id))
             .await
     }
 
@@ -265,7 +283,11 @@ impl DatabaseApi<'_> {
         link_type: LinkType,
     ) -> Result<PersistedOntologyMetadata, InsertionError> {
         self.store
-            .create_link_type(link_type, self.account_id, self.account_id)
+            .create_link_type(
+                link_type,
+                OwnedById::new(self.account_id),
+                CreatedById::new(self.account_id),
+            )
             .await
     }
 
@@ -295,7 +317,7 @@ impl DatabaseApi<'_> {
         link_type: LinkType,
     ) -> Result<PersistedOntologyMetadata, UpdateError> {
         self.store
-            .update_link_type(link_type, self.account_id)
+            .update_link_type(link_type, UpdatedById::new(self.account_id))
             .await
     }
 
@@ -309,24 +331,29 @@ impl DatabaseApi<'_> {
             .create_entity(
                 entity,
                 entity_type_id,
-                self.account_id,
+                OwnedById::new(self.account_id),
                 entity_id,
-                self.account_id,
+                CreatedById::new(self.account_id),
             )
             .await
     }
 
     pub async fn get_entity(&mut self, entity_id: EntityId) -> Result<PersistedEntity, QueryError> {
-        Ok(self
+        let vertex = self
             .store
             .get_entity(&StructuralQuery {
                 expression: Expression::for_latest_entity_id(entity_id),
                 graph_resolve_depths: GraphResolveDepths::zeroed(),
             })
             .await?
-            .pop()
-            .expect("no entity found")
-            .entity)
+            .vertices
+            .remove(&GraphElementIdentifier::KnowledgeGraphElementId(entity_id))
+            .expect("no entity found");
+
+        match vertex {
+            Vertex::Entity(persisted_entity) => Ok(persisted_entity),
+            _ => unreachable!(),
+        }
     }
 
     pub async fn update_entity(
@@ -336,7 +363,12 @@ impl DatabaseApi<'_> {
         entity_type_id: VersionedUri,
     ) -> Result<PersistedEntityMetadata, UpdateError> {
         self.store
-            .update_entity(entity_id, entity, entity_type_id, self.account_id)
+            .update_entity(
+                entity_id,
+                entity,
+                entity_type_id,
+                UpdatedById::new(self.account_id),
+            )
             .await
     }
 
@@ -348,7 +380,11 @@ impl DatabaseApi<'_> {
     ) -> Result<(), InsertionError> {
         let link = Link::new(source_entity_id, target_entity_id, link_type_id, None);
         self.store
-            .create_link(&link, self.account_id, self.account_id)
+            .create_link(
+                &link,
+                OwnedById::new(self.account_id),
+                CreatedById::new(self.account_id),
+            )
             .await
     }
 
@@ -366,7 +402,11 @@ impl DatabaseApi<'_> {
             Some(index),
         );
         self.store
-            .create_link(&link, self.account_id, self.account_id)
+            .create_link(
+                &link,
+                OwnedById::new(self.account_id),
+                CreatedById::new(self.account_id),
+            )
             .await
     }
 
@@ -439,7 +479,9 @@ impl DatabaseApi<'_> {
         link_type_id: VersionedUri,
     ) -> Result<(), LinkRemovalError> {
         let link = Link::new(source_entity_id, target_entity_id, link_type_id, None);
-        self.store.remove_link(&link, self.account_id).await
+        self.store
+            .remove_link(&link, RemovedById::new(self.account_id))
+            .await
     }
 }