+++ packages/hash/api/src/model/knowledge/entity.model.ts
@@ -1,9 +1,9 @@
 import { ApolloError } from "apollo-server-errors";
-
 import {
   PersistedEntity,
   GraphApi,
   StructuralQuery,
+  Vertex,
 } from "@hashintel/hash-graph-client";
 
 import {
@@ -307,7 +307,7 @@ export default class {
     query: object,
     options?: Omit<Partial<StructuralQuery>, "query">,
   ): Promise<EntityModel[]> {
-    const { data: entityRootedSubgraphs } = await graphApi.getEntitiesByQuery({
+    const { data: subgraph } = await graphApi.getEntitiesByQuery({
       query,
       graphResolveDepths: {
         dataTypeResolveDepth:
@@ -325,9 +325,13 @@ export default class {
     });
 
     return await Promise.all(
-      entityRootedSubgraphs.map(({ entity }) =>
-        EntityModel.fromPersistedEntity(graphApi, entity),
-      ),
+      subgraph.roots.map((entityId) => {
+        const entityVertex = subgraph.vertices[entityId] as Extract<
+          Vertex,
+          { kind: "entity" }
+        >;
+        return EntityModel.fromPersistedEntity(graphApi, entityVertex.inner);
+      }),
     );
   }