+++ packages/hash/frontend/src/blocks/page/Comments/CommentBlock.tsx
@@ -3,8 +3,8 @@ import {
   ReactNode,
   useEffect,
   useMemo,
-  useRef,
   useState,
+  useCallback,
 } from "react";
 import { Box, Collapse, Tooltip, Typography } from "@mui/material";
 import {
@@ -85,14 +85,13 @@ export const CommentBlock: FunctionComponent<CommentProps> = ({
 }) => {
   const { entityId, hasText, author, textUpdatedAt } = comment;
 
-  const ref = useRef<HTMLDivElement>(null);
+  const [container, setContainer] = useState<HTMLDivElement | null>(null);
   const [collapsed, setCollapsed] = useState(true);
   const [shouldCollapse, setShouldCollapse] = useState(false);
   const [editable, setEditable] = useState(false);
   const [inputValue, setInputValue] = useState<TextToken[]>(hasText);
   const [deleteConfirmationDialogOpen, setDeleteConfirmationDialogOpen] =
     useState(false);
-  const [isAnimating, setIsAnimating] = useState(false);
 
   const { user } = useUser();
   const [updateCommentText, { loading: updateCommentTextLoading }] =
@@ -136,7 +135,6 @@ export const CommentBlock: FunctionComponent<CommentProps> = ({
   const resetCommentText = () => {
     setInputValue(hasText);
     setEditable(false);
-    setIsAnimating(true);
   };
 
   const handleEditComment = async () => {
@@ -146,9 +144,14 @@ export const CommentBlock: FunctionComponent<CommentProps> = ({
     }
   };
 
+  const onLineCountChange = useCallback(
+    (lines: number) => !editable && setShouldCollapse(lines > 2),
+    [editable, setShouldCollapse],
+  );
+
   return (
     <Box
-      ref={ref}
+      ref={(ref: HTMLDivElement) => setContainer(ref)}
       sx={{
         display: "flex",
         flexDirection: "column",
@@ -157,6 +160,10 @@ export const CommentBlock: FunctionComponent<CommentProps> = ({
         borderTop: ({ palette }) => `1px solid ${palette.gray[20]}`,
         minHeight: deleteConfirmationDialogOpen ? 140 : 0,
         transition: ({ transitions }) => transitions.create("min-height"),
+
+        ":first-of-type": {
+          borderTopWidth: 0,
+        },
       }}
     >
       <Box display="flex" justifyContent="space-between">
@@ -251,12 +258,9 @@ export const CommentBlock: FunctionComponent<CommentProps> = ({
               borderColor: palette.blue[60],
             },
           })}
-          onTransitionEnd={() => setIsAnimating(false)}
         >
           <CommentTextField
-            onLineCountChange={(lines) =>
-              !editable && !isAnimating && setShouldCollapse(lines > 2)
-            }
+            onLineCountChange={onLineCountChange}
             value={inputValue}
             className={`${styles.Comment__TextField} ${
               editable
@@ -339,7 +343,6 @@ export const CommentBlock: FunctionComponent<CommentProps> = ({
               setEditable(!editable);
               setCollapsed(false);
               setShouldCollapse(false);
-              setIsAnimating(true);
               commentMenuPopupState.close();
             }}
           />
@@ -365,7 +368,7 @@ export const CommentBlock: FunctionComponent<CommentProps> = ({
       </CommentBlockMenu>
 
       <CommentBlockDeleteConfirmationDialog
-        container={ref.current}
+        container={container}
         open={deleteConfirmationDialogOpen}
         loading={deleteCommentLoading}
         onDelete={async () => await deleteComment(entityId)}