+++ packages/hash/frontend/src/components/hooks/blockProtocolFunctions/knowledge/knowledge-shim.ts
@@ -13,12 +13,13 @@ import {
 } from "@blockprotocol/graph";
 import {
   PersistedLink,
-  Subgraph,
   UnknownPersistedEntity,
 } from "../../../../graphql/apiTypes.gen";
+import { Subgraph } from "../../../../lib/subgraph";
 
 export type KnowledgeCallbacks = {
   getEntity: GetEntityMessageCallback;
+  aggregateEntities: AggregateEntitiesMessageCallback;
   updateEntity: UpdateEntityMessageCallback;
 };
 
@@ -39,9 +40,9 @@ type BaseEntity = Omit<
   | "entityType"
 >;
 
-type Entity = BaseEntity & {
+/** @todo: remove this when we start returning links in the subgraph - https://app.asana.com/0/0/1203214689883095/f */
+export type Entity = BaseEntity & {
   links: Link[];
-  entityTypeRootedSubgraph: Subgraph;
 };
 
 type Link = Omit<PersistedLink, "sourceEntity" | "targetEntity"> & {
@@ -54,7 +55,15 @@ export type GetEntityRequest = Pick<EntityResponse, "entityId">;
 export type GetEntityMessageCallback = MessageCallback<
   GetEntityRequest,
   null,
-  EntityResponse,
+  Subgraph,
+  ReadOrModifyResourceError
+>;
+
+export type AggregateEntitiesRequest = {};
+export type AggregateEntitiesMessageCallback = MessageCallback<
+  AggregateEntitiesRequest,
+  null,
+  Subgraph,
   ReadOrModifyResourceError
 >;