+++ packages/hash/frontend/src/components/hooks/blockProtocolFunctions/ontology/useBlockProtocolGetDataType.ts
@@ -7,6 +7,7 @@ import {
 } from "../../../../graphql/apiTypes.gen";
 import { getDataTypeQuery } from "../../../../graphql/queries/ontology/data-type.queries";
 import { GetDataTypeMessageCallback } from "./ontology-types-shim";
+import { Subgraph } from "../../../../lib/subgraph";
 
 export const useBlockProtocolGetDataType = (): {
   getDataType: GetDataTypeMessageCallback;
@@ -53,7 +54,11 @@ export const useBlockProtocolGetDataType = (): {
       }
 
       return {
-        data: response.data.getDataType,
+        /**
+         * @todo: remove this when we start returning links in the subgraph
+         *   https://app.asana.com/0/0/1203214689883095/f
+         */
+        data: response.data.getDataType as Subgraph,
       };
     },
     [getFn],